requirements.txt
aiogram==3.10.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
aiosqlite==0.20.0
loguru==0.7.2
pydantic==2.8.2
